==> 2_data_cleaning.ipynb <==
import os
import pandas as pd
import re
import polars as pl


# Path to your extracted CSV files
data_path = r"C:\Users\anbun\Desktop\Portfolio projects\Google Data Analytics Capstone\Data\12_Months_data"

# Output file path
output_file = os.path.join(data_path, "combined_cyclistic_data.csv")


#Combining the 12 months data together from oct-24 to sep-25
# List all CSV files in the directory
csv_files = [f for f in os.listdir(data_path) if f.endswith(".csv")]

print(f"Found {len(csv_files)} CSV files.")

combined_df = pd.DataFrame()

for file in csv_files:
    # Extract yyyymm from filename (e.g., 202304-divvy-tripdata.csv → 202304)
    month_tag = file[:6]  # first 6 characters
    year = month_tag[:4]
    month = month_tag[4:6]
    month_year = f"{month}-{year}"  # mm-yyyy format

    file_path = os.path.join(data_path, file)
    print(f"Processing {file} ...")

    # Read CSV file
    df = pl.read_csv(file_path)

    # Add month_year column
    df["month_year"] = month_year

    # Append to combined dataframe
    combined_df = pl.concat([combined_df, df], ignore_index=True)


combined_df.shape


# understanding the data and the columns data types
combined_df.info()


# checking for missing values
combined_df.isnull().sum()


combined_df["rideable_type"].value_counts(dropna=False)


combined_df["member_casual"].value_counts(dropna=False)


combined_df["month_year"].value_counts(dropna=False)


combined_df['ride_id_length'] = combined_df['ride_id'].astype(str).apply(len)


print(combined_df['ride_id_length'].value_counts().sort_index())


# --- CONVERT DATES TO DATETIME ---
combined_df['started_at'] = pd.to_datetime(combined_df['started_at'], errors='coerce')
combined_df['ended_at'] = pd.to_datetime(combined_df['ended_at'], errors='coerce')


combined_df.info()


# --- CALCULATE RIDE TIME IN MINUTES ---
combined_df['ride_time_min'] = (combined_df['ended_at'] - combined_df['started_at']).dt.total_seconds() / 60


unusual_rides = combined_df[(combined_df['ride_time_min'] <= 1) | (combined_df['ride_time_min'] >= 1440)]

print(f"Found {len(unusual_rides)} rides with unusual lengths.\n")


combined_df.shape


# --- DROP ROWS ---
combined_df = combined_df.drop(combined_df[(combined_df['ride_time_min'] <= 1) | (combined_df['ride_time_min'] >= 1440)].index)

# --- DISPLAY TOTAL ROWS AFTER ---
print(f"Rows remaining after deleting unusual rides: {len(combined_df)}")


len(combined_df)


# --- CHECK FOR EMPTY (NULL or BLANK) START STATION NAMES ---
missing_start = combined_df[combined_df['start_station_name'].isna() | (combined_df['start_station_name'].astype(str).str.strip() == '')]
print(f"Rows with missing start_station_name: {len(missing_start)}")


# --- CHECK FOR EMPTY (NULL or BLANK) END STATION NAMES ---
missing_end = combined_df[combined_df['end_station_name'].isna() | (combined_df['end_station_name'].astype(str).str.strip() == '')]
print(f"Rows with missing end_station_name: {len(missing_end)}")


# --- CHECK MISSING STATIONS ---
missing_start = combined_df[combined_df['start_station_name'].isna() | (combined_df['start_station_name'].astype(str).str.strip() == '')]
missing_end = combined_df[combined_df['end_station_name'].isna() | (combined_df['end_station_name'].astype(str).str.strip() == '')]
missing_start_id = combined_df[combined_df['start_station_id'].isna() | (combined_df['start_station_id'].astype(str).str.strip() == '')]
missing_end_id = combined_df[combined_df['end_station_id'].isna() | (combined_df['end_station_id'].astype(str).str.strip() == '')]


# --- ANALYZE rideable_type DISTRIBUTION FOR MISSING STATIONS ---
print("Distribution of RIDEABLE TYPE for rides with missing START station:")
print("-" * 70)
print(missing_start['rideable_type'].value_counts(dropna=False))


print("\nDistribution of RIDEABLE TYPE for rides with missing END station:")
print("-" * 70)
print(missing_end['rideable_type'].value_counts(dropna=False))


print("\nDistribution of RIDEABLE TYPE for rides with missing END station:")
print("-" * 70)
print(missing_start_id['rideable_type'].value_counts(dropna=False))


print("\nDistribution of RIDEABLE TYPE for rides with missing END station:")
print("-" * 70)
print(missing_end_id['rideable_type'].value_counts(dropna=False))


unusual_start_end = combined_df[(combined_df['start_station_name'].isna()) | (combined_df['end_station_name'].isna())| (combined_df['start_station_id'].isna()) | (combined_df['end_station_id'].isna())]

print(f"Found {len(unusual_start_end)} rides with unusual start or end stations.\n")


# --- DROP ROWS ---
combined_df = combined_df.drop(combined_df[(combined_df['start_station_name'].isna()) | (combined_df['end_station_name'].isna())| (combined_df['start_station_id'].isna()) | (combined_df['end_station_id'].isna())].index)

# --- DISPLAY TOTAL ROWS AFTER ---
print(f"Rows remaining after deleting unusual start or end stations: {len(combined_df)}")


combined_df['start_station_name'].nunique()


combined_df['end_station_name'].nunique()


combined_df['start_station_id'].nunique()


combined_df['end_station_id'].nunique()


combined_df.isnull().sum()


# --- GROUP BY STATION NAME AND COUNT DISTINCT IDs ---
station_id_counts = (
    combined_df.groupby('start_station_name')['start_station_id']
      .nunique()  # equivalent to COUNT(DISTINCT start_station_id)
      .reset_index(name='distinct_ids')
)


stations_with_multiple_ids = station_id_counts[station_id_counts['distinct_ids'] > 1]
stations_with_multiple_ids


# for name in stations_with_multiple_ids['start_station_name']:
#     ids = combined_df.loc[combined_df['start_station_name'] == name, 'start_station_id'].unique()
#     print(f"{name}: {ids}")


# --- GROUP BY STATION NAME AND COUNT DISTINCT IDs ---
station_id_counts = (
    combined_df.groupby('end_station_name')['end_station_id']
      .nunique()  # equivalent to COUNT(DISTINCT start_station_id)
      .reset_index(name='distinct_ids')
)


stations_with_multiple_ids = station_id_counts[station_id_counts['distinct_ids'] > 1]
stations_with_multiple_ids


print("🔍 Checking for NULL or empty values in start_lat:")
start_lat_summary = (
    combined_df['start_lat']
    .value_counts(dropna=False)
    .sort_index()
)


print("\n🔍 Checking for NULL or empty values in end_lat:")
end_lat_summary = (
    combined_df['end_lat']
    .value_counts(dropna=False)
    .sort_index()
)


# Identify rows with null, blank, or zero end_lat
condition = (
    combined_df['end_lat'].isna() |
    (combined_df['end_lat'].astype(str).str.strip() == '') |
    (combined_df['end_lat'].astype(str).str.strip() == '0')|
    (combined_df['end_lat'] == 0)
)

print(f"\n🚫 Rows with invalid end_lat: {condition.sum()}")


combined_df.shape


combined_df.head()


combined_df.info()


combined_df['month'] = combined_df['started_at'].dt.strftime('%b')


combined_df['weekday'] = combined_df['started_at'].dt.strftime('%a') 


combined_df.info()


combined_df


season_map ={    
    'Dec':'Winter','Jan':'Winter','Feb':'Winter',
    'Mar':'Spring','Apr':'Spring','May':'Spring',
    'Jun':'Summer','Jul':'Summer','Aug':'Summer',
    'Sep':'Fall','Oct':'Fall','Nov':'Fall'
    }

combined_df['season']= combined_df['month'].map(season_map)

season_order = ['Winter','Spring','Summer','Fall']
combined_df['season']= pd.Categorical(combined_df['season'],
                                      categories= season_order,
                                      ordered =True)


combined_df.head()


combined_df['hour'] = (
    combined_df['started_at']
      .dt.strftime('%I %p')              # e.g., '01 AM'
      .str.replace(r'^0', '', regex=True) # drop leading zero -> '1 AM'
      .str.lower()                        # '1 am'
)

hours_order = (
    [f'{h} am' for h in [12,1,2,3,4,5,6,7,8,9,10,11]] +
    [f'{h} pm' for h in [12,1,2,3,4,5,6,7,8,9,10,11]]
)
combined_df['hour'] = combined_df['hour'].astype(pd.CategoricalDtype(categories=hours_order, ordered=True))


# --- SAVE THE COMBINED FILE ---
combined_df.to_csv(output_file, index=False)
